requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-lifecycle==0.9.1
django-quill-editor==0.1.22
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
urlman==2.0.1
